+++ tket/src/Circuit/ToffoliBox.cpp
@@ -27,9 +27,11 @@
 namespace tket {
 
 ToffoliBox::ToffoliBox(
-    const state_perm_t &permutation, const OpType &rotation_axis)
+    const state_perm_t &permutation, const ToffoliBoxSynthStrat &strat,
+    const OpType &rotation_axis)
     : Box(OpType::ToffoliBox),
       permutation_(permutation),
+      strat_(strat),
       rotation_axis_(rotation_axis) {
   // we need to check every element has the same size, and every bitstrings
   // appears twice
@@ -69,6 +71,7 @@ ToffoliBox::ToffoliBox(
 ToffoliBox::ToffoliBox(const ToffoliBox &other)
     : Box(other),
       permutation_(other.permutation_),
+      strat_(other.strat_),
       rotation_axis_(other.rotation_axis_) {}
 
 Op_ptr ToffoliBox::dagger() const {
@@ -76,7 +79,7 @@ Op_ptr ToffoliBox::dagger() const {
   for (auto it = permutation_.begin(); it != permutation_.end(); it++) {
     reverse_perm.insert({it->second, it->first});
   }
-  return std::make_shared<ToffoliBox>(reverse_perm, rotation_axis_);
+  return std::make_shared<ToffoliBox>(reverse_perm, strat_, rotation_axis_);
 }
 
 Op_ptr ToffoliBox::transpose() const { return dagger(); }
@@ -90,6 +93,8 @@ op_signature_t ToffoliBox::get_signature() const {
 state_perm_t ToffoliBox::get_permutation() const { return permutation_; }
 
 OpType ToffoliBox::get_rotation_axis() const { return rotation_axis_; }
+
+ToffoliBoxSynthStrat ToffoliBox::get_strat() const { return strat_; }
 
 typedef boost::adjacency_list<boost::vecS, boost::vecS, boost::undirectedS>
     cube_graph_t;
@@ -308,9 +313,17 @@ static Circuit permute(
   return circ;
 }
 
+static Circuit gen_circuit_using_toffoli_gates(
+    state_perm_t &perm, unsigned n_qubits);
+
 void ToffoliBox::generate_circuit() const {
   unsigned n_qubits = (unsigned)permutation_.begin()->first.size();
   state_perm_t perm(permutation_);
+  if (this->strat_ == ToffoliBoxSynthStrat::Cycle) {
+    circ_ = std::make_shared<Circuit>(
+        gen_circuit_using_toffoli_gates(perm, n_qubits));
+    return;
+  }
   // fill the permutation with identities
   if (perm.size() != (1ULL << n_qubits)) {
     for (unsigned long long i = 0; i < (1ULL << n_qubits); i++) {
@@ -324,10 +337,295 @@ void ToffoliBox::generate_circuit() const {
       permute(perm, n_qubits, get_op_ptr(rotation_axis_, 1)));
 }
 
+// Old ToffoliBox decomposition method
+
+struct transposition_t {
+  std::vector<bool> first;
+  std::vector<bool> middle;
+  std::vector<bool> last;
+};
+typedef std::vector<std::vector<bool>> cycle_permutation_t;
+typedef std::vector<transposition_t> cycle_transposition_t;
+typedef std::vector<std::pair<std::vector<bool>, unsigned>> gray_code_t;
+
+static unsigned get_hamming_distance(
+    const std::vector<bool> &a, const std::vector<bool> &b) {
+  if (a.size() != b.size()) {
+    throw std::invalid_argument("Bitstrings must have identical size.");
+  }
+  unsigned counter = 0;
+  for (unsigned i = 0; i < a.size(); i++) {
+    if (a[i] != b[i]) {
+      ++counter;
+    }
+  }
+  return counter;
+}
+
+static cycle_transposition_t cycle_to_transposition(cycle_permutation_t cycle) {
+  /**
+   * A cycle can start at any element
+   * A transposition for a cycle can always be constructed by pairing the
+   * starting element with the others in cycle order
+   * This also gives opportunities to produce gray codes with matching elements
+   * that can be cancelled
+   *
+   * For each element in a cycle, produce a sequence of transpositions and
+   * compare total Hamming distance Return the transposition with smallest
+   * Hamming distance
+   *
+   */
+  cycle_transposition_t best_transposition;
+  unsigned best_hamming_distance = 0;
+
+  for (unsigned i = 0; i < cycle.size(); i++) {
+    unsigned accumulated_hamming_distance = 0;
+    cycle_transposition_t transposition;
+    for (unsigned j = 1; j < cycle.size(); j++) {
+      transposition.push_back({cycle[0], cycle[j], cycle[0]});
+      accumulated_hamming_distance += get_hamming_distance(cycle[0], cycle[j]);
+    }
+    if (best_transposition.empty() ||
+        accumulated_hamming_distance < best_hamming_distance) {
+      best_transposition = transposition;
+      best_hamming_distance = accumulated_hamming_distance;
+    }
+    std::rotate(cycle.begin(), cycle.begin() + 1, cycle.end());
+  }
+  return best_transposition;
+}
+
+static std::vector<cycle_transposition_t> get_transpositions(
+    const std::set<cycle_permutation_t> &cycles) {
+  std::vector<cycle_transposition_t> transpositions;
+  for (const cycle_permutation_t &cycle : cycles) {
+    // each cycle is costed via the Hamming distance to reduce the number of
+    // operations
+    transpositions.push_back(cycle_to_transposition(cycle));
+  }
+  return transpositions;
+}
+
+static cycle_transposition_t merge_cycles(
+    std::vector<cycle_transposition_t> &cycle_transpositions) {
+  cycle_transposition_t return_transposition;
+  for (unsigned k = 0; k < cycle_transpositions.size(); k++) {
+    cycle_transposition_t cycle = cycle_transpositions[k];
+    unsigned i = 0, j = 1;
+    while (j < cycle.size()) {
+      transposition_t transposition_i = cycle[i];
+      transposition_t transposition_j = cycle[j];
+
+      std::vector<bool> transposition_j_first = transposition_j.first;
+      std::vector<bool> transposition_i_last = transposition_i.last;
+      std::vector<bool> transposition_i_first = transposition_i.first;
+
+      TKET_ASSERT(transposition_i_last == transposition_j.first);
+      std::vector<bool> i_middle = transposition_i.middle;
+      std::vector<bool> j_middle = transposition_j.middle;
+
+      TKET_ASSERT(i_middle.size() == transposition_i.last.size());
+      TKET_ASSERT(j_middle.size() == transposition_i.last.size());
+      // if a transposition has already been reduced, still need to make sure we
+      // uncompute it
+      if (transposition_i_first != transposition_i_last) {
+        unsigned middle_last_distance =
+            get_hamming_distance(i_middle, transposition_i_last);
+        unsigned middle_first_distance =
+            get_hamming_distance(i_middle, transposition_i_first);
+        // this => the reduced transposition is on a good gray code between the
+        // new "first" and target
+        if (middle_first_distance < middle_last_distance &&
+            middle_first_distance > 1) {
+          transposition_i_last = transposition_i_first;
+          std::vector<bool> starting_point = transposition_i_last;
+          for (unsigned k = 0; k < i_middle.size(); k++) {
+            if (i_middle[k] == j_middle[k] &&
+                get_hamming_distance(starting_point, i_middle) > 1) {
+              starting_point[k] = i_middle[k];
+            }
+          }
+
+          cycle_transpositions[k][i].last = starting_point;
+          cycle_transpositions[k][j].first = starting_point;
+        }
+      } else {  // else in this case just find any good transposition
+        std::vector<bool> starting_point = transposition_i_last;
+        for (unsigned k = 0; k < i_middle.size(); k++) {
+          if (i_middle[k] == j_middle[k] &&
+              get_hamming_distance(starting_point, i_middle) > 1) {
+            starting_point[k] = i_middle[k];
+          }
+        }
+
+        cycle[i].last = starting_point;
+        cycle[j].first = starting_point;
+      }
+
+      ++i;
+      ++j;
+    }
+
+    return_transposition.insert(
+        return_transposition.end(), cycle.begin(), cycle.end());
+  }
+
+  return return_transposition;
+}
+
+static gray_code_t transposition_to_gray_code(
+    const transposition_t &transposition) {
+  unsigned first_middle_hamming_distance =
+      get_hamming_distance(transposition.first, transposition.middle);
+  unsigned middle_last_hamming_distance =
+      get_hamming_distance(transposition.middle, transposition.last);
+  gray_code_t all_gray_code_entries;
+  // => that some optimisation is done to middle_last, so must go via
+  // transposition.last bitstring to allow proper cancellation
+  // At some point transposition.first == transposition.last
+  // If transposition.last != transposition.first & the hamming distance
+  // between transposition.last and transposition.middle is smaller than
+  // transposition.first and transposition.middle,
+  // then we need to make sure that the gray code between transposition.first
+  // and transposition.middle goes via transposition.last, such that the
+  // eventual gray code between transposition.middle and transposition.last
+  // uncomputes this we can assume that the remaining
+  // transposition.last->transposition.first gray code would be cancelled out if
+  // added
+  std::vector<bool> initial = transposition.first;
+  if (middle_last_hamming_distance < first_middle_hamming_distance) {
+    // get bitstrings for first -> last
+    for (unsigned i = 0; i < transposition.first.size(); i++) {
+      if (transposition.first[i] != transposition.last[i]) {
+        initial[i] = !initial[i];
+        all_gray_code_entries.push_back({initial, i});
+      }
+    }
+  }
+  // with the right middle bitstring now guaranteed, go from this bitstring to
+  // middle
+  std::vector<bool> bitstring = initial;
+  for (unsigned i = 0; i < transposition.first.size(); i++) {
+    if (initial[i] != transposition.middle[i]) {
+      bitstring[i] = !bitstring[i];
+      all_gray_code_entries.push_back({bitstring, i});
+    }
+  }
+  // now do the last->middle in reverse to guarantee right
+  // gray code path is taken
+  initial = transposition.last;
+  gray_code_t reverse_gray_code_entries;
+  // as before, implies some optimisation has been completed on
+  // first one
+  // thus make sure that gray code goes via transposition.first
+  // to make applied permutation right
+  if (first_middle_hamming_distance < middle_last_hamming_distance) {
+    // get bitstrings for first -> last
+    for (unsigned i = 0; i < transposition.first.size(); i++) {
+      if (initial[i] != transposition.first[i]) {
+        initial[i] = !initial[i];
+        reverse_gray_code_entries.push_back({initial, i});
+      }
+    }
+  }
+  // and then go from bitstring to middle
+  bitstring = initial;
+  for (unsigned i = 0; i < transposition.middle.size(); i++) {
+    if (transposition.middle[i] != initial[i]) {
+      bitstring[i] = !bitstring[i];
+      reverse_gray_code_entries.push_back({bitstring, i});
+    }
+  }
+  // don't want to add transformation for reaching final -> so pop_back
+  if (!reverse_gray_code_entries.empty()) {
+    reverse_gray_code_entries.pop_back();
+  }
+  all_gray_code_entries.insert(
+      all_gray_code_entries.end(), reverse_gray_code_entries.rbegin(),
+      reverse_gray_code_entries.rend());
+  return all_gray_code_entries;
+}
+
+Circuit get_bitstring_circuit(
+    const std::vector<bool> &bitstring, const unsigned &target,
+    const unsigned &n_qubits) {
+  // flip qubits that need to be state 0
+  Circuit x_circuit(n_qubits);
+  std::vector<unsigned> cnx_args;
+  for (unsigned i = 0; i < n_qubits; i++) {
+    if (i != target) {
+      if (!bitstring[i]) {
+        x_circuit.add_op<unsigned>(OpType::X, {i});
+      }
+      cnx_args.push_back(i);
+    }
+  }
+  cnx_args.push_back(target);
+  TKET_ASSERT(cnx_args.size() == n_qubits);
+
+  Circuit return_circuit(n_qubits);
+  return_circuit.append(x_circuit);
+  return_circuit.add_op<unsigned>(OpType::CnX, cnx_args);
+  return_circuit.append(x_circuit);
+  return return_circuit;
+}
+
+static Circuit gen_circuit_using_toffoli_gates(
+    state_perm_t &perm, unsigned n_qubits) {
+  // Convert passed permutation to cycles
+  std::set<cycle_permutation_t> cycles;
+  while (!perm.empty()) {
+    auto it = perm.begin();
+    cycle_permutation_t cycle = {it->first};
+    it = perm.find(it->second);
+    while (it->first != cycle[0]) {
+      cycle.push_back(it->first);
+      it = perm.find(it->second);
+    }
+    if (cycle.size() > 1) {
+      cycles.insert(cycle);
+    }
+    for (const std::vector<bool> &bitstring : cycle) {
+      perm.erase(bitstring);
+    }
+  }
+
+  // This decomposition is as described on page 191, section 4.5.2 "Single
+  // qubit and CNOT gates are universal" of Nielsen & Chuang
+  std::vector<cycle_transposition_t> cycle_transpositions =
+      get_transpositions(cycles);
+
+  // optionally, order the transpositions and cycles to allow gate
+  // cancellation
+  cycle_transposition_t ordered_transpositions;
+  ordered_transpositions = merge_cycles(cycle_transpositions);
+
+  if (ordered_transpositions.empty()) {
+    return Circuit(n_qubits);
+  }
+
+  // Now we have ordered transpositions, produced front->middle and
+  // middle->back gray codes for each transposition and add to circuit
+  Circuit circ(n_qubits);
+  for (const transposition_t &transposition : ordered_transpositions) {
+    TKET_ASSERT(transposition.first.size() == n_qubits);
+    TKET_ASSERT(transposition.middle.size() == n_qubits);
+    TKET_ASSERT(transposition.last.size() == n_qubits);
+    gray_code_t all_gray_code_entries =
+        transposition_to_gray_code(transposition);
+    for (const std::pair<std::vector<bool>, unsigned> &entry :
+         all_gray_code_entries) {
+      circ.append(get_bitstring_circuit(entry.first, entry.second, n_qubits));
+    }
+  }
+  return circ;
+}
+
 nlohmann::json ToffoliBox::to_json(const Op_ptr &op) {
   const auto &box = static_cast<const ToffoliBox &>(*op);
   nlohmann::json j = core_box_json(box);
   j["permutation"] = box.get_permutation();
+  j["strat"] = box.get_strat();
   j["rotation_axis"] = box.get_rotation_axis();
   return j;
 }
@@ -335,6 +633,7 @@ nlohmann::json ToffoliBox::to_json(const Op_ptr &op) {
 Op_ptr ToffoliBox::from_json(const nlohmann::json &j) {
   ToffoliBox box = ToffoliBox(
       j.at("permutation").get<state_perm_t>(),
+      j.at("strat").get<ToffoliBoxSynthStrat>(),
       j.at("rotation_axis").get<OpType>());
   return set_box_id(
       box,